consumer/requirements.txt
hnswlib==0.2
numpy==1.16.0
paho-mqtt==1.4.0
psycopg2==2.7.6.1
pybind11==2.2.4
